+++ test/pretty/resource/after/PatternInstanceOf.java
@@ -1,0 +1,7 @@
+public class PatternInstanceOf {
+	public void foo(Object o) {
+		if (o instanceof String y) {
+			System.out.println(y);
+		}
+	}
+}