+++ test/pretty/resource/before/PatternInstanceOf.java
@@ -1,0 +1,8 @@
+// version 14:
+public class PatternInstanceOf {
+	public void foo(Object o) {
+		if (o instanceof String y) {
+			System.out.println(y);
+		}
+	}
+}